// src/context/enums.ts
export const enum ChatActionType {
  SET_TABS = 'SET_TABS',
  CREATE_TAB = 'CREATE_TAB',
  REMOVE_TAB = 'REMOVE_TAB',
  SET_ACTIVE_TAB = 'SET_ACTIVE_TAB',
  SET_CONVERSATIONS = 'SET_CONVERSATIONS',
  ADD_CONVERSATION = 'ADD_CONVERSATION',
  SET_CURRENT_MESSAGE = 'SET_CURRENT_MESSAGE',
  CLEAR_CURRENT_MESSAGE = 'CLEAR_CURRENT_MESSAGE',
  SET_LOADING = 'SET_LOADING',
  SET_STREAMING = 'SET_STREAMING',
  UPDATE_STREAMING_MESSAGE = 'UPDATE_STREAMING_MESSAGE',
  SET_ERROR = 'SET_ERROR',
  CLEAR_MESSAGES = 'CLEAR_MESSAGES',
  SET_EDITING_MESSAGE_ID = 'SET_EDITING_MESSAGE_ID',
  UPDATE_EDITING_MESSAGE = 'UPDATE_EDITING_MESSAGE',
}

export const enum EChatUserType {
  USER = 'user',
  ASSISTANT = 'assistant',
  SYSTEM = 'system',
}